fix: [BE] 출금 데드락 예방·입력 검증 400 매핑·정렬 개선 (#60)

+++ backend/src/main/java/com/gighub/wallet/controller/WalletController.java
@@ -115,7 +115,7 @@ public class WalletController {
                 .maxAmount(maxAmount)
                 .keyword(trimmedKeyword)
                 .sort(sort)
-                .offset(page * size)
+                .offset((long)page * size)
                 .size(size)
                 .build();
